Concernsで共通化しつつbookmarkを拾うアクションを作っておく
