+++ app/helpers/bookmarks_helper.rb
@@ -1,0 +1,2 @@
+module BookmarksHelper
+end
