+++ config/routes.rb
@@ -18,6 +18,7 @@ Rails.application.routes.draw do
   post '/follow/:user_id', to: 'users#follow_user', as: :follow_user
 
   resources :users, param: :name, only: %i[show edit update]
+  resources :bookmarks, only: [:show]
 
   resources :tasks
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
